Restructure markdown editor components.
