+++ coral/src/app/components/documentation/utils/topic-documentation-helper.ts
@@ -3,6 +3,9 @@ import rehypeParse from "rehype-parse";
 import rehypeSanitize from "rehype-sanitize";
 import rehypeRemark from "rehype-remark";
 import remarkStringify from "remark-stringify";
+import remarkParse from "remark-parse";
+import remarkRehype from "remark-rehype";
+import rehypeStringify from "rehype-stringify";
 
 async function createMarkdown(stringifiedHtml: string) {
   const result = await unified()
@@ -14,4 +17,15 @@ async function createMarkdown(stringifiedHtml: string) {
   return String(result);
 }
 
-export { createMarkdown };
+async function createStringifiedHtml(markdownText: string) {
+  const result = await unified()
+    .use(remarkParse)
+    .use(remarkRehype)
+    .use(rehypeSanitize)
+    .use(rehypeStringify)
+    .process(markdownText);
+
+  return String(result);
+}
+
+export { createMarkdown, createStringifiedHtml };
